api/requirements.txt
Flask
Flask-RESTful
PyJWT
pymongo
Jinja2